requirements.txt
Pillow
torch
numpy
opencv-python
sounddevice
scipy
whisper
omegaconf
transformers
huggingface-hub
unittest
ultralytics
